+++ freppledb/output/views/demand.py
@@ -27,7 +27,7 @@ from django.utils.encoding import force_text
 from freppledb.boot import getAttributeFields
 from freppledb.input.models import Demand, Item
 from freppledb.input.models import ManufacturingOrder, PurchaseOrder, DistributionOrder
-from freppledb.common.report import GridPivot, GridFieldText
+from freppledb.common.report import GridPivot, GridFieldText, GridFieldNumber
 from freppledb.common.report import GridFieldCurrency, GridFieldLastModified
 
 
@@ -64,6 +64,8 @@ class OverviewReport(GridPivot):
             initially_hidden=True,
         ),
         GridFieldCurrency("cost", title=_("cost"), initially_hidden=True),
+        GridFieldNumber("volume", title=_("volume"), initially_hidden=True),
+        GridFieldNumber("weight", title=_("weight"), initially_hidden=True),
         GridFieldText("source", title=_("source"), initially_hidden=True),
         GridFieldLastModified("lastmodified", initially_hidden=True),
     )
@@ -150,7 +152,7 @@ class OverviewReport(GridPivot):
         query = """
           select
           parent.name, parent.description, parent.category, parent.subcategory,
-          parent.owner_id, parent.cost, parent.source, parent.lastmodified,
+          parent.owner_id, parent.cost, parent.volume, parent.weight, parent.source, parent.lastmodified,
           %s
           d.bucket,
           d.startdate,
@@ -195,7 +197,8 @@ class OverviewReport(GridPivot):
                        ) d
           group by
             parent.name, parent.description, parent.category, parent.subcategory,
-            parent.owner_id, parent.cost, parent.source, parent.lastmodified, parent.lft, parent.rght,
+            parent.owner_id, parent.cost, parent.volume, parent.weight, parent.source,
+            parent.lastmodified, parent.lft, parent.rght,
             %s
             d.bucket, d.startdate, d.enddate
           order by %s, d.startdate
@@ -233,8 +236,10 @@ class OverviewReport(GridPivot):
                         "subcategory": row[3],
                         "owner": row[4],
                         "cost": row[5],
-                        "source": row[6],
-                        "lastmodified": row[7],
+                        "volume": row[6],
+                        "weight": row[7],
+                        "source": row[8],
+                        "lastmodified": row[9],
                         "bucket": row[numfields - 6],
                         "startdate": row[numfields - 5].date(),
                         "enddate": row[numfields - 4].date(),
@@ -243,7 +248,7 @@ class OverviewReport(GridPivot):
                         "reasons": json.dumps(row[numfields - 1]),
                         "backlog": backlog,
                     }
-                    idx = 8
+                    idx = 10
                     for f in getAttributeFields(Item):
                         res[f.field_name] = row[idx]
                         idx += 1